added on cascade for deletion to allow deletion of objects with foreign keys, added if-statements the deletion methods in service classes instead, to limit what deletion actions are allowed

+++ src/main/java/com/backendContextAssignment1/plantSwap_postgres/Services/PlantService.java
@@ -47,7 +47,9 @@ public class PlantService {
     }
 
     public void deletePlantById(Long id) {
-        validatePlantIdAndReturnPlant(id);
+        if (validatePlantIdAndReturnPlant(id).getAvailabilityStatus() == PlantAvailabilityStatusEnum.RESERVED) {
+            throw new UnsupportedOperationException("plants with pending transactions cannot be deleted");
+        }
         plantRepository.deleteById(id);
     }
